fixes along NanoBob suggestions
